requirements.txt
Flask==3.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
psycopg2-binary>=2.9.9
python-dotenv==1.0.0
flask-cors==4.0.0
google-cloud-storage==2.10.0
google-cloud-documentai==2.20.1
gunicorn==21.2.0
openpyxl>=3.1.0
pytest==7.4.3

